requirements.txt
requests
discord
psycopg2
bs4
python-dotenv